Move SegmentedText to core root
