+++ src/components/Text/Annotated/core/AnnotatedText.tsx
@@ -8,7 +8,7 @@ import {
   Segment,
   TextOffsets,
 } from "./AnnotationModel.ts";
-import { SegmentedText } from "./inline/SegmentedText.tsx";
+import { SegmentedText } from "./SegmentedText.tsx";
 import { createContext, FC, PropsWithChildren, ReactNode } from "react";
 import { Any } from "../../../../utils/Any.ts";
 import { BlockSchema } from "./block";
